+++ src/main/java/com/ifonly/museagent/scheduler/AliveCheckScheduler.java
@@ -118,5 +118,4 @@ public class AliveCheckScheduler implements SchedulingConfigurer {
       scheduleExecutorService.purgeExpiredTrash();
     }
   }
-
 }